inital

+++ enegine_domainnet_celeba.py
@@ -67,7 +67,7 @@ def train_distill(args, student, teacher, optimizer, dataloader):
             loss = Mse(stud_logits, teach_logits)
         elif args.dis_loss=='direct_label_sample':
             sampler = torch.distributions.bernoulli.Bernoulli(Sig(teach_hid))
-            teach_label = sampler.sample().float()
+            teach_label = sampler.sample().int().float()
             loss = Bce(stud_hid,teach_label)
         else:
             print('dis_loss must be cesample, argmax, or mse')
